refactor: use vanilla-extract in dialog component
