+++ src/components/dialog/Dialog.css.ts
@@ -1,0 +1,52 @@
+import { ComplexStyleRule, keyframes } from "@vanilla-extract/css";
+import { recipe, RecipeVariants } from "@vanilla-extract/recipes";
+import { color } from "../../theme/color.css";
+import { config } from "../../theme/config.css";
+import { DefaultReset } from "../reset.css";
+import { ContainerColor } from "../types";
+
+const getVariant = (variant: ContainerColor): ComplexStyleRule => ({
+  backgroundColor: color[variant].Container,
+  color: color[variant].OnContainer,
+  border: `${config.borderWidth.B300} solid ${color[variant].ContainerLine}`,
+});
+
+const DialogOpenAnime = keyframes({
+  "0%": {
+    transform: "translateY(5px)",
+  },
+  "100%": {
+    transform: "translateY(0)",
+  },
+});
+
+export const Dialog = recipe({
+  base: [
+    DefaultReset,
+    {
+      borderRadius: config.radii.R400,
+      boxShadow: config.shadow.E400,
+      width: "100%",
+      maxWidth: config.size.DialogWidth,
+      overflow: "hidden",
+      animation: `${DialogOpenAnime} 200ms`,
+    },
+  ],
+  variants: {
+    variant: {
+      Background: getVariant("Background"),
+      Surface: getVariant("Surface"),
+      SurfaceVariant: getVariant("SurfaceVariant"),
+      Primary: getVariant("Primary"),
+      Secondary: getVariant("Secondary"),
+      Success: getVariant("Success"),
+      Warning: getVariant("Warning"),
+      Critical: getVariant("Critical"),
+    },
+  },
+  defaultVariants: {
+    variant: "Surface",
+  },
+});
+
+export type DialogVariants = RecipeVariants<typeof Dialog>;
