+++ src/components/dialog/Dialog.tsx
@@ -1,42 +1,10 @@
-import { CSS, keyframes, styled } from "../../config";
-import { ContainerColor } from "../types";
+import classNames from "classnames";
+import React from "react";
+import { as } from "../as";
+import * as css from "./Dialog.css";
 
-const getVariant = (variant: ContainerColor): CSS => ({
-  backgroundColor: `$${variant}Container`,
-  color: `$On${variant}Container`,
-  border: `$borderWidths$300 solid $colors$${variant}ContainerLine`,
-});
-
-const DialogOpenAnime = keyframes({
-  "0%": {
-    transform: "translateY(5px)",
-  },
-  "100%": {
-    transform: "translateY(0)",
-  },
-});
-
-export const Dialog = styled("div", {
-  borderRadius: "$400",
-  boxShadow: "$E400",
-  width: "100%",
-  maxWidth: "$DialogWidth",
-  overflow: "hidden",
-  animation: `${DialogOpenAnime} 200ms`,
-
-  variants: {
-    variant: {
-      Background: getVariant("Background"),
-      Surface: getVariant("Surface"),
-      SurfaceVariant: getVariant("SurfaceVariant"),
-      Primary: getVariant("Primary"),
-      Secondary: getVariant("Secondary"),
-      Success: getVariant("Success"),
-      Warning: getVariant("Warning"),
-      Critical: getVariant("Critical"),
-    },
-  },
-  defaultVariants: {
-    variant: "Surface",
-  },
-});
+export const Dialog = as<"div", css.DialogVariants>(
+  ({ as: AsDialog = "div", className, variant, ...props }, ref) => (
+    <AsDialog className={classNames(css.Dialog({ variant }), className)} {...props} ref={ref} />
+  )
+);
